Change balance properties from float to double for precision
Updated the data type of several balance-related properties in
the `CSFloatAndSteamBalance` class from `float` to `double` to
improve precision. Affected properties include:
- `CSFloatUsableBalanceInPounds`
- `CSFloatTotalBalanceInPounds`
- `CSFloatPendingBalanceInPounds`
- `SteamUsableBalanceInPounds`
- `SteamPendingBalanceInPounds`
- `SteamTotalBalanceInPounds`

+++ Omnipotent/Services/CS2ArbitrageBot/CS2ArbitrageBotLabs/Scanalytics.cs
@@ -109,12 +109,12 @@ namespace Omnipotent.Services.CS2ArbitrageBot.CS2ArbitrageBotLabs
 
         public class CSFloatAndSteamBalance
         {
-            public float CSFloatUsableBalanceInPounds { get; set; }
-            public float CSFloatTotalBalanceInPounds { get; set; }
-            public float CSFloatPendingBalanceInPounds { get; set; }
-            public float SteamUsableBalanceInPounds { get; set; }
-            public float SteamPendingBalanceInPounds { get; set; }
-            public float SteamTotalBalanceInPounds { get; set; }
+            public double CSFloatUsableBalanceInPounds { get; set; }
+            public double CSFloatTotalBalanceInPounds { get; set; }
+            public double CSFloatPendingBalanceInPounds { get; set; }
+            public double SteamUsableBalanceInPounds { get; set; }
+            public double SteamPendingBalanceInPounds { get; set; }
+            public double SteamTotalBalanceInPounds { get; set; }
 
             public DateTime DateTimeOfBalanceRecord { get; set; }
             public CSFloatWrapper.Statistics CSFloatProfileStatistics { get; set; }
